+++ library/std/src/sync/poison/once.rs
